backend/requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9
djangorestframework>=3.15,<3.16
django-filter>=24.1
drf-spectacular>=0.27
django-cors-headers>=4.3
dj-database-url>=2.1
python-dotenv>=1.0
